#592 - Refactoring common types and helper methods out of Staffing-related classes (#596)
