+++ backend/Api/Common/Types/BookingType.cs
@@ -1,0 +1,11 @@
+namespace Api.Common.Types;
+
+public enum BookingType
+{
+	Offer,
+	Booking,
+	PlannedAbsence,
+	Vacation,
+	Available,
+	NotStartedOrQuit
+}
