+++ backend/Api/Common/Types/UpdateDepartmentReadModel.cs
@@ -1,0 +1,3 @@
+namespace Api.Common.Types;
+
+public record UpdateDepartmentReadModel(string Id, string Name);
