+++ backend/Api/StaffingController/StaffingReadModel.cs
@@ -1,3 +1,4 @@
+using Api.Common.Types;
 using Core.Consultants;
 using Core.Weeks;
 
@@ -31,23 +32,6 @@ public record StaffingReadModel(
             bookings,
             detailedBookings,
             IsOccupied
-        )
-    {
-    }
-}
-
-public record UpdateDepartmentReadModel(
-    string Id,
-    string Name);
-
-public record CompetenceReadModel(
-    string Id,
-    string Name)
-{
-    public CompetenceReadModel(Competence competence)
-        : this(
-            competence.Id,
-            competence.Name
         )
     {
     }
@@ -91,23 +75,4 @@ public record WeeklyBookingReadModel(
     double TotalOverbooking,
     double TotalNotStartedOrQuit);
 
-public record BookingDetails(
-    string ProjectName,
-    BookingType Type,
-    string CustomerName,
-    int ProjectId,
-    bool ExcludeFromBilling = false,
-    bool IsBillable = false,
-    DateTime? EndDateAgreement = null);
-
 public record WeeklyHours(int Week, double Hours);
-
-public enum BookingType
-{
-    Offer,
-    Booking,
-    PlannedAbsence,
-    Vacation,
-    Available,
-    NotStartedOrQuit
-}
